pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartmarket"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "Django>=5,<6",
  "djangorestframework>=3.15",
  "django-filter>=24.2",
  "drf-spectacular>=0.27",
  "psycopg2-binary>=2.9"
]

[project.optional-dependencies]
dev = [
  "django-debug-toolbar>=4.3",
  "pytest",
  "pytest-django",
  "black",
  "ruff"
]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.black]
line-length = 200
target-version = ["py311"]

[tool.ruff]
line-length = 200
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I","UP","B"]
ignore = ["D","F403","F405"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py"]
